requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
Django==5.1.2
djangorestframework==3.15.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
jiter==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==0.28.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
